services/profile-api/requirements.txt
Flask==1.1.2
PyYAML==5.3.1
flask-cors==3.0.8
mysql-connector==2.2.9